Checks and modifications of the propagation time step
If the field interaction is off, there is no need to have a substep for
propation. The exponential propagator is exact, hence, the propagation
step can be set to printing step. This is now reflected in the code. If
V_int=0 or field=0, the code will set dt=print_time. If the interaction
is on, substeps are necessary to make the propagator valid.